move flat-options.js from lib/config/ to lib/utils/
