+++ lib/npm.js
@@ -141,7 +141,7 @@ const npm = module.exports = new class extends EventEmitter {
         this.log.warn('using --force', 'Recommended protections disabled.')
       }
       if (!er && !this[_flatOptions]) {
-        this[_flatOptions] = require('./config/flat-options.js')(this)
+        this[_flatOptions] = require('./utils/flat-options.js')(this)
         process.env.npm_command = this.command
       }
       process.emit('timeEnd', 'npm:load')
