+++ lib/utils/flat-options.js
